Allowing finer control on claiming logic
This patch adds a `ShouldClaimPendingMessage` callback to the
subscriber's config, to allow users to decide whether a message that's
been pending for more than `MaxIdleTime` should actually be claimed or
not.

My use case for this is that I have a type of async tasks for which the
time needed to process one task can vary wildly, anywhere from 1 second
to several minutes, and it's impossible to know ahead of time how long
it will take. So to avoid long-running instances to be considered dead
and be endlessly claimed back & forth, I would have to have a
`MaxIdleTime` of several minutes.
However, that would also meant that it would take that long to detect
dead workers, which is also not acceptable for my use case.

This new config option solves this problem by allowing me to have a
short `MaxIdleTime`, but then leverage an external heartbeat mechanism
in my `ShouldClaimPendingMessage` callback to decide if a worker is
actually dead, or just busy with a long-running instance of the task.

Added tests on the claim behaviour.

Signed-off-by: Jean Rouge <rougej+github@gmail.com>