ack invalid messages (#38)
After a while messages are being lost from the queue because of
constraints in the queue size.

```
error -- redis: nil
xp -- {
        Consumer: dispatcher-7f7484b54f-njdpq
        ID: 1722875908017-0
        Idle: 1020707000000
        RetryCount: 1427
      }
```

What is happening is that the message exists in the PEL but has no
content beacues it was deleted. Acking the message removes it from the
PEL so it won't be claimed again in the same consumer group.